requirements.txt
cffi==1.11.2
cryptography==1.7.2
Django==1.11.7
django-rest-framework==0.1.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
idna==2.6
jwt==0.5.2
pyasn1==0.3.7
pycparser==2.18
PyJWT==1.5.3
pytz==2017.3
six==1.11.0
